용용

+++ guad_backend/src/main/java/auction/guad/vo/RequestMileageVo.java
@@ -9,7 +9,7 @@ public class RequestMileageVo {
 
 	private String email;
 	private Date chargeDate;
-	private int chargeAmount;
+	private long chargeAmount;
 	private String chargeMethod;
 	
 	//
